smartdocs_pro/backend/requirements.txt
Flask
flask-cors
requests
pdfplumber
python-docx
pytesseract
python-dotenv
sqlite-utils
groq
streamlit
pdf2image
torch
transformers
sentencepiece
langchain
langchain-community
faiss-cpu
pypdf
sentence-transformers
tiktoken